Commit del ejercicio 4 y 5 tp 1 PV
Descripción: se coloco una opción de control en ejercicio 4 y finalizo
el ejercicio 5.
Autor: Abel Peñaloza

+++ tp1_AbelPenhaloza/src/ar/edu/unju/fi/ejercicio1/Main.java
@@ -10,7 +10,7 @@ public class Main {
 		int num4=20;
 		int num5=9;
 		//Calcula el promedio - se coloca 5.0 para asegurar obtener un numero decimal.
-		double promedio = (num1+num2+num3+num4+num5)/5.0;
+		double promedio = (double)(num1+num2+num3+num4+num5)/5;
 		//Mostrar el resultado por consola
 		System.out.println("El promedio es: "+ promedio);
